requirements.txt
asgiref~=3.5.2
channels~=4.0.0
constantly~=15.1.0
cryptography~=38.0.3
daphne~=4.0.0
Django~=3.2.21
hyperlink~=21.0.0
django-resized~=1.0.2
Pillow~=9.2.0
djangorestframework~=3.14.0
dj-database-url~=2.1.0
gunicorn~=20.1.0
flake8~=5.0.0
django-cors-headers~=3.14.0
